test(sass): move `BranchMatcher` to `branch.py`
