+++ setup.py
@@ -23,6 +23,7 @@ def enable_mypyc(dist: Distribution) -> None:
             'reprospect/test/sass/composite.py',
             'reprospect/test/sass/controlflow/block.py',
             'reprospect/test/sass/instruction/address.py',
+            'reprospect/test/sass/instruction/branch.py',
             'reprospect/test/sass/instruction/constant.py',
             'reprospect/test/sass/instruction/half.py',
             'reprospect/test/sass/instruction/instruction.py',
